requirements.txt
flask
requests
openai
python-dotenv
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1